Assignment 2 Web Engineering Widget Dashboard
In this assignments we are required to make a single page website which
shows a dashboard in the start. The dashboard should show different
interactive plugins to the user. e.g.

Calendar
Clock
Weather

The User should be able to click on any of these plugins and get a
detailed view of the same. e.g. if a user clicks on Calendar, they
should get a Monthly Calendar view on the page, or if a user clicks on
Clock, they should get a Clock on the page and not the dashboard
anymore.
On any particular plugin page, the user should be able to go back to the
home page.